src/main/java/ru/job4j/it/BackwardArrayIt.java
package ru.job4j.it;

import java.util.Iterator;
import java.util.NoSuchElementException;

public class BackwardArrayIt implements Iterator<Integer> {

    private final int[] data;
    private int point = 0;

    public BackwardArrayIt(int[] data) {
        this.data = data;
        for (int i = 0; i < data.length; i++) {
            point = i;
        }
    }

    @Override
    public boolean hasNext() {
        return point >= 0 && data.length != 0;
    }

    @Override
    public Integer next() {
        if (!hasNext()) {
            throw new NoSuchElementException();
        }
        return data[point--];
    }
}