src/test/java/ru/job4j/it/BackwardArrayItTest.java
package ru.job4j.it;


import static org.hamcrest.MatcherAssert.assertThat;
import static org.hamcrest.Matchers.is;
import static org.junit.Assert.assertFalse;

import java.util.NoSuchElementException;
import org.junit.Test;

public class BackwardArrayItTest {

    @Test

    public void whenReadSequenceAndHasNextThenFalse() {
        BackwardArrayIt it = new BackwardArrayIt(new int[]{1, 2, 3, 4});
        assertThat(it.next(), is(4));
        assertThat(it.next(), is(3));
        assertThat(it.next(), is(2));
        assertThat(it.next(), is(1));
        assertFalse(it.hasNext());
    }

    @Test
    public void whenMultiCallHasNextThenTrue() {
        BackwardArrayIt it = new BackwardArrayIt(
            new int[]{1, 2, 3}
        );
        assertThat(it.hasNext(), is(true));
        assertThat(it.hasNext(), is(true));
    }

    @Test
    public void whenReadSequence() {
        BackwardArrayIt it = new BackwardArrayIt(
            new int[]{1, 2, 3}
        );
        assertThat(it.next(), is(3));
        assertThat(it.next(), is(2));
        assertThat(it.next(), is(1));
    }

    @Test(expected = NoSuchElementException.class)
    public void whenNextFromEmpty() {
        BackwardArrayIt it = new BackwardArrayIt(
            new int[]{}
        );
        it.next();
    }
}